+++ source/common/modules/markdown-editor/statusbar/info-fields.ts
@@ -1,0 +1,96 @@
+/**
+ * @ignore
+ * BEGIN HEADER
+ *
+ * Contains:        Info Statusbar Items
+ * CVM-Role:        View
+ * Maintainer:      Hendrik Erz
+ * License:         GNU GPL v3
+*
+ * Description:     This file defines a set of info statusbar items
+ *
+ * END HEADER
+ */
+
+import { type EditorState } from '@codemirror/state'
+import { type EditorView } from '@codemirror/view'
+import { trans } from '@common/i18n-renderer'
+import localiseNumber from '@common/util/localise-number'
+import { type StatusbarItem } from '.'
+import { countField } from '../plugins/statistics-fields'
+import { configField } from '../util/configuration'
+
+/**
+ * Displays the cursor position
+ *
+ * @param   {EditorState}    state  The EditorState
+ * @param   {EditorView}     view   The EditorView
+ *
+ * @return  {StatusbarItem}         Returns the element
+ */
+export function cursorStatus (state: EditorState, view: EditorView): StatusbarItem|null {
+  const mainOffset = state.selection.main.head
+  const line = state.doc.lineAt(mainOffset)
+  return {
+    content: `${line.number}:${mainOffset - line.from + 1}`
+  }
+}
+
+/**
+ * Displays the word count, if applicable
+ *
+ * @param   {EditorState}    state  The EditorState
+ * @param   {EditorView}     view   The EditorView
+ *
+ * @return  {StatusbarItem}         Returns the element or null
+ */
+export function wordcountStatus (state: EditorState, view: EditorView): StatusbarItem|null {
+  const counter = state.field(countField, false)
+  if (counter === undefined) {
+    return null
+  } else {
+    return {
+      content: trans('%s words', localiseNumber(counter.words))
+    }
+  }
+}
+
+/**
+ * Displays the character count, if applicable
+ *
+ * @param   {EditorState}    state  The EditorState
+ * @param   {EditorView}     view   The EditorView
+ *
+ * @return  {StatusbarItem}         Returns the element or null
+ */
+export function charcountStatus (state: EditorState, view: EditorView): StatusbarItem|null {
+  const counter = state.field(countField, false)
+  if (counter === undefined) {
+    return null
+  } else {
+    return {
+      content: trans('%s characters', localiseNumber(counter.chars))
+    }
+  }
+}
+
+/**
+ * Displays an input mode indication, if applicable
+ *
+ * @param   {EditorState}    state  The EditorState
+ * @param   {EditorView}     view   The EditorView
+ *
+ * @return  {StatusbarItem}         Returns the element or null
+ */
+export function inputModeStatus (state: EditorState, view: EditorView): StatusbarItem|null {
+  const config = state.field(configField, false)
+  if (config === undefined) {
+    return null
+  } else if (config.inputMode !== 'default') {
+    return {
+      content: 'Mode: ' + (config.inputMode === 'vim' ? 'Vim' : 'Emacs')
+    }
+  } else {
+    return null
+  }
+}