+++ source/win-main/store/index.ts
@@ -14,14 +14,13 @@
  * END HEADER
  */
 
-import { StoreOptions, createStore as baseCreateStore, Store } from 'vuex'
-import { InjectionKey } from 'vue'
-import { ColoredTag } from '@providers/tags'
-import { SearchResultWrapper } from '@dts/common/search'
-import { RelatedFile } from '@dts/renderer/misc'
+import { createStore as baseCreateStore, type StoreOptions, type Store } from 'vuex'
+import { type InjectionKey } from 'vue'
+import { type ColoredTag } from '@providers/tags'
+import type { SearchResultWrapper } from '@dts/common/search'
 import locateByPath from '@providers/fsal/util/locate-by-path'
 import configToArrayMapper from './config-to-array'
-import { BranchNodeJSON, LeafNodeJSON, OpenDocument } from '@dts/common/documents'
+import type { BranchNodeJSON, LeafNodeJSON, OpenDocument } from '@dts/common/documents'
 
 // Import Mutations
 import addToFiletreeMutation from './mutations/add-to-filetree'
@@ -33,11 +32,10 @@ import documentTreeMutation from './mutations/document-tree'
 // Import Actions
 import filetreeUpdateAction from './actions/filetree-update'
 import updateOpenDirectoryAction from './actions/update-open-directory'
-import updateRelatedFilesAction from './actions/update-related-files'
 import updateBibliographyAction from './actions/update-bibliography'
 import documentTreeUpdateAction from './actions/document-tree-update'
-import { AnyDescriptor, DirDescriptor, MaybeRootDescriptor } from '@dts/common/fsal'
-import { WritingTarget } from '@providers/targets'
+import type { AnyDescriptor, DirDescriptor, MaybeRootDescriptor } from '@dts/common/fsal'
+import { type WritingTarget } from '@providers/targets'
 import updateSnippetsAction from './actions/update-snippets'
 
 const ipcRenderer = window.ipc
@@ -78,15 +76,6 @@ export interface ZettlrState {
    */
   selectedDirectory: DirDescriptor|null
   activeFile: null
-  /**
-   * This property contains all leaf IDs on which the readability mode is
-   * currently active
-   */
-  readabilityModeActive: string[]
-  /**
-   * Files which are in some way related to the currently active file
-   */
-  relatedFiles: RelatedFile[]
   /**
    * Contains coloured tags that can be managed in the tag manager
    */
@@ -158,11 +147,9 @@ function getConfig (): StoreOptions<ZettlrState> {
         paneData: [],
         fileTree: [],
         lastFiletreeUpdate: 0,
-        readabilityModeActive: [],
         activeFile: null,
         uncollapsedDirectories: [],
         selectedDirectory: null,
-        relatedFiles: [],
         colouredTags: [],
         writingTargets: [],
         config: configToArrayMapper(window.config.get()),
@@ -233,28 +220,11 @@ function getConfig (): StoreOptions<ZettlrState> {
           state.uncollapsedDirectories = oldUncollapsed
         }
       },
-      addReadabilityActiveLeaf (state, leaf) {
-        if (!state.readabilityModeActive.includes(leaf)) {
-          state.readabilityModeActive.push(leaf)
-        }
-      },
-      removeReadabilityActiveLeaf (state, leaf) {
-        const idx = state.readabilityModeActive.indexOf(leaf)
-        if (idx > -1) {
-          state.readabilityModeActive.splice(idx, 1)
-        }
-      },
       updateConfig: function (state, option) {
         state.config[option] = window.config.get(option)
       },
       lastFiletreeUpdate: function (state, payload) {
         state.lastFiletreeUpdate = payload
-      },
-      updateRelatedFiles: function (state, relatedFiles: RelatedFile[]) {
-        // Make sure we're only updating if something has changed.
-        if (JSON.stringify(relatedFiles) !== JSON.stringify(state.relatedFiles)) {
-          state.relatedFiles = relatedFiles
-        }
       },
       updateModifiedFiles: function (state, modifiedDocuments: string[]) {
         state.modifiedDocuments = modifiedDocuments
@@ -299,10 +269,7 @@ function getConfig (): StoreOptions<ZettlrState> {
       updateOpenDirectory: updateOpenDirectoryAction,
       lastLeafId: async function (ctx, lastLeafId: string) {
         ctx.commit('lastLeafId', lastLeafId)
-        // Update the related files
-        await ctx.dispatch('updateRelatedFiles')
       },
-      updateRelatedFiles: updateRelatedFilesAction,
       updateBibliography: updateBibliographyAction,
       documentTree: documentTreeUpdateAction,
       updateSnippets: updateSnippetsAction,